Enemy damage sending to server and receiving by player

+++ ShooterClient/Assets/Scripts/Character.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public abstract class Character : MonoBehaviour
 {
-    [field: SerializeField] public float MaxHealth { get; protected set; } = 10;
+    [field: SerializeField] public int MaxHealth { get; protected set; } = 10;
     [field: SerializeField] public float Speed { get; protected set; } = 2f;
 
     public Vector3 Velocity { get; protected set; }
